+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 2015-3-3
 --------
 - Added Form\UploadProfilePictureForm test
+- Added Options\ModuleOptions test